refactored DrinkCard to be in DrinkCard directory
